# BE/swagger-api-docs.yaml
openapi: 3.0.0
info:
  title: Dashboard CAPSTONE 12 BRIN
  description: Official API documentation for Dashboard CAPSTONE 12 BRIN
  version: 1.0.0
...
paths:
  /api/v1/users:
    get:
      tags:
        - Users
      summary: Fetch all users
      parameters:
        - schema:
            type: string
            example: john
            description: Searched through names and emails (Case Insensitive)
          required: false
          description: Searched through names and emails (Case Insensitive)
          name: search
          in: query
        - schema:
            type: number
            minimum: 1
            example: 1
          required: false
          name: roleId
          in: query
        - schema:
            type: number
            minimum: 1
            maximum: 100
            default: 20
            example: 20
          required: false
          name: limit
          in: query
        - schema:
            type: number
            nullable: true
            minimum: 0
            example: 0
          required: false
          name: offset
          in: query
      responses:
        "200":
          description: Users fetched
    post:
      tags:
        - Users
      summary: Create user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                roleId:
                  type: number
                  minimum: 1
                  example: 1
                name:
                  type: string
                  minLength: 1
                  example: John Doe
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  minLength: 6
                  example: secret123
                confirmPassword:
                  type: string
                  minLength: 6
                  example: secret123
              required:
                - roleId
                - name
                - email
                - password
                - confirmPassword
              title: CreateUserRequest
      responses:
        "201":
          description: User created
        "400":
          description: Invalid input
        "409":
          description: User already exists
...
  api/v1/identifications:
    get:
      tags:
        - Identifications
      summary: Fetch all identifications
      parameters:
        - schema:
            type: string
            example: Test
            description: Searched through attributes (Case Insensitive)
          required: false
          description: Searched through attributes (Case Insensitive)
          name: search
          in: query
        - schema:
            type: number
            minimum: 1
            maximum: 100
            default: 20
            example: 20
          required: false
          name: limit
          in: query
        - schema:
            type: number
            nullable: true
            minimum: 0
            example: 0
          required: false
          name: offset
          in: query
        - schema:
            type: number
            minimum: 1
            example: 1
          required: false
          name: plantId
          in: query
        - schema:
            type: boolean
            example: true
          required: false
          name: isSuccess
          in: query
      responses:
        "200":
          description: Identifications fetched
...